Update index.php
So, I'm messing around with editing the text on the home page, but am not sure about headers, grid sizes, etc.  I guess I just need to know if I should be changing the <h1>, etc things or just entering the text I want somewhere.  I started to edit header stuff a bit, but don't want to mess it up so stopped and just entered text (not even sure if it is entered in the right area ~ there was a "main" and an "article" section and I am not sure which I should be working in).  Sorry for the questions/confusion, but I think I will figure it out soon with a bit more practice.  Also, I am not sure where the headers (tabs?) are.  I know I saw them on the mock up site you showed me (Home, Bids&Fees, Logistics, ..)