Remove H5BP index.html file.
I think I had placed it in the root OLD directory when I was making the KK27 site to keep it safe and out of the way.
Now that it's value has been ported into the PARTIAL php scripts, it isn't necessary to keep it around anymore.
Note that this may impact future pulls from the H5BP upstream (as those changes will likely be to this index.html file.