Remove unnecessary $regform existence check.
This condition was a leftover when this JS was included in the MAIN.js
file, when it only made sense to run it if form#registration existed on
the page.

The JS now lives in the same page as form#registration, so no need the
check (or the extra indent ;) ).