Set late player fee.
I had written all the special sauce to handle the edge cases where the page is loaded before the deadline, but then the player was added after the deadline.

However, I hadn't handled the IMPORTANT case of when the page is loaded after the deadline.  Now when the page is loaded after the deadline, the player fee is appropriately set.