Revamp Contacts page.
Now utilize PHP OOP methodology to easily modify layout and content of the Kommittee.
The KommitteePosition class defines a certain position on the Kommittee.
Each KommitteePosition takes at least one KommitteeMember object (which can be created independently of the KommitteePosition).  That is, it is a matter of moving the KommitteeMember object to another KommitteePosition if someone changes jobs.